Move hidden columns hooks into useColumns
